# backend/app/routers/users.py
"""User endpoints (api-contracts Users). Org Admin manages tenant users;
Super Admin creates other Super Admins."""
from __future__ import annotations

from fastapi import APIRouter, Depends, File, Query, UploadFile
from sqlalchemy.ext.asyncio import AsyncSession

from app.dependencies import Principal, db_session, require_roles
from app.middleware.errors import AppError
from app.schemas.user import (
    BulkCreateParticipantsRequest,
    BulkCreateParticipantsResult,
    BulkParticipantRow,
    CreateSuperAdminRequest,
    CreateUserRequest,
    UpdateUserRequest,
    UserOut,
)
from app.services import user_service as svc
from app.utils.csv_import import parse_participant_csv

router = APIRouter(tags=["Users"])
_org_admin = require_roles("ORG_ADMIN")
_super_admin = require_roles("SUPER_ADMIN")


def _require_tenant(principal: Principal) -> str:
    if principal.tenant_id is None:
        raise AppError(403, "FORBIDDEN", "Operation requires a tenant-scoped user")
    return principal.tenant_id


@router.post("/users", response_model=UserOut, status_code=201)
async def create_user(
    body: CreateUserRequest,
    principal: Principal = Depends(_org_admin),
    session: AsyncSession = Depends(db_session),
) -> UserOut:
    tenant_id = _require_tenant(principal)
    return UserOut.model_validate(await svc.create_user(session, tenant_id, body))


@router.post("/users/bulk", response_model=BulkCreateParticipantsResult)
async def bulk_create_participants(
    file: UploadFile = File(...),
    principal: Principal = Depends(_org_admin),
    session: AsyncSession = Depends(db_session),
) -> BulkCreateParticipantsResult:
    """Bulk-import PARTICIPANT accounts from a CSV file (F5, FR-3a).

    The CSV must contain a header row with columns: email, first_name, last_name.
    Malformed rows and duplicates (already in the tenant or repeated within the
    file) are SKIPPED with a reason; valid new rows are CREATED with a generated
    one-time password returned for out-of-band distribution.
    """
    tenant_id = _require_tenant(principal)
    if file.content_type not in ("text/csv", "application/vnd.ms-excel", None):
        raise AppError(415, "UNSUPPORTED_MEDIA_TYPE", "Only CSV files are supported")
    try:
        text = (await file.read()).decode("utf-8")
    except UnicodeDecodeError as exc:
        raise AppError(422, "INVALID_CSV", "CSV must be UTF-8 encoded") from exc

    rows = parse_participant_csv(text)
    body = BulkCreateParticipantsRequest(participants=[BulkParticipantRow(**r) for r in rows])
    return await svc.bulk_create_participants(session, tenant_id, body)


@router.get("/users", response_model=list[UserOut])
async def list_users(
    role: str | None = Query(None),
    status: str | None = Query(None),
    limit: int = Query(50, ge=1, le=200),
    principal: Principal = Depends(_org_admin),
    session: AsyncSession = Depends(db_session),
) -> list[UserOut]:
    tenant_id = _require_tenant(principal)
    users = await svc.list_users(session, tenant_id, role=role, status=status, limit=limit)
    return [UserOut.model_validate(u) for u in users]


@router.get("/users/{user_id}", response_model=UserOut)
async def get_user(
    user_id: str,
    principal: Principal = Depends(_org_admin),
    session: AsyncSession = Depends(db_session),
) -> UserOut:
    tenant_id = _require_tenant(principal)
    return UserOut.model_validate(await svc.get_user(session, tenant_id, user_id))


@router.patch("/users/{user_id}", response_model=UserOut)
async def update_user(
    user_id: str,
    body: UpdateUserRequest,
    principal: Principal = Depends(_org_admin),
    session: AsyncSession = Depends(db_session),
) -> UserOut:
    tenant_id = _require_tenant(principal)
    return UserOut.model_validate(await svc.update_user(session, tenant_id, user_id, body))


@router.post("/super-admins", response_model=UserOut, status_code=201, tags=["Users"])
async def create_super_admin(
    body: CreateSuperAdminRequest,
    _: Principal = Depends(_super_admin),
    session: AsyncSession = Depends(db_session),
) -> UserOut:
    return UserOut.model_validate(await svc.create_super_admin(session, body))