# docs/spec/api-contracts.yaml
openapi: 3.0.3
info:
  title: ContestForge Core Contest Engine API
  version: 1.0.0
  description: >
    Multi-tenant live contest engine. REST endpoints cover tenancy, user
    management, contest authoring, lifecycle, registration, live-runtime
    controls, results, notifications, and audit. Live question delivery,
    answer submission, leaderboard, and elimination events flow over a
    WebSocket channel (documented in api-contracts.md §WebSocket).

    Authentication: every endpoint except /auth/login, /auth/refresh, /health,
    and /ready requires a JWT bearer token whose claims include the caller's
    role and tenant_id. Cross-tenant access is rejected with 403.

    Onboarding: the first Super Admin is bootstrapped by an SQL script and is
    not exposed in this API. A Super Admin creates an organization (which
    provisions its initial Org Admin). An Org Admin creates co-Org-Admins,
    Moderators, and Participants via POST /users. There is no public self-signup.
servers:
  - url: https://api.contestforge.example/v1
    description: Placeholder; region/host finalized in /neutron:plan
security:
  - bearerAuth: []
...
paths:
...
  /users:
    post:
      tags: [Users]
      summary: Create a tenant user (Org Admin). Role limited to ORG_ADMIN, MODERATOR, PARTICIPANT.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserRequest' }
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Email already exists in this tenant
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
    get:
      tags: [Users]
      summary: List users in the caller's tenant (Org Admin)
      parameters:
        - in: query
          name: role
          schema: { type: string, enum: [ORG_ADMIN, MODERATOR, PARTICIPANT] }
        - in: query
          name: status
          schema: { type: string, enum: [ACTIVE, DISABLED] }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200':
          description: Users
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/User' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /users/bulk:
    post:
      tags: [Users]
      summary: Bulk-import Participants (Org Admin). Creates PARTICIPANT accounts from a CSV/JSON list.
      description: >
        Creates PARTICIPANT accounts in the caller's tenant from an array of
        { email, first_name, last_name }. Generates a one-time password per new
        user (v1 has no email delivery) returned to the Org Admin for
        out-of-band distribution. Partial success: duplicates or invalid rows
        are skipped and reported per row; valid rows still succeed. Role is
        always PARTICIPANT (other roles use POST /users).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BulkCreateParticipantsRequest' }
          text/csv:
            schema:
              type: string
              format: binary
              description: CSV with header `email,first_name,last_name`
      responses:
        '200':
          description: Import processed (see per-row results)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BulkCreateParticipantsResult' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /users/{userId}:
    parameters:
      - { $ref: '#/components/parameters/UserId' }
    get:
      tags: [Users]
      summary: Get a user in the caller's tenant
      responses:
        '200':
          description: User
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Users]
      summary: Update a user's name or status (Org Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name: { type: string }
                last_name: { type: string }
                status: { type: string, enum: [ACTIVE, DISABLED] }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /super-admins:
    post:
      tags: [Users]
      summary: Create a new Super Admin. Existing Super Admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSuperAdminRequest' }
      responses:
        '201':
          description: Super Admin created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Email already exists for a Super Admin
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }

  # ───────────────────────────── Organizations ─────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    UserId:
      in: path
      name: userId
      required: true
      schema: { type: string, format: uuid }
...
    Limit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      in: query
      name: cursor
      schema: { type: string }
      description: Opaque cursor returned by the previous page
  responses:
...
    Forbidden:
      description: Role or tenant not permitted
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found in caller's tenant
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    UnprocessableEntity:
      description: Semantic validation error (e.g. elimination block missing rules)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
    Page:
      type: object
      description: Pagination envelope; concrete `items` defined per endpoint
      properties:
        next_cursor: { type: string, nullable: true }
        has_more: { type: boolean }
...
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        tenant_id: { type: string, format: uuid, nullable: true }
        email: { type: string, format: email }
        first_name: { type: string }
        last_name: { type: string }
        role: { type: string, enum: [SUPER_ADMIN, ORG_ADMIN, MODERATOR, PARTICIPANT] }
        status: { type: string, enum: [ACTIVE, DISABLED] }
        created_at: { type: string, format: date-time }
    CreateUserRequest:
      type: object
      required: [email, first_name, last_name, role, password]
      properties:
        email: { type: string, format: email }
        first_name: { type: string }
        last_name: { type: string }
        role:
          type: string
          enum: [ORG_ADMIN, MODERATOR, PARTICIPANT]
          description: SUPER_ADMIN is not allowed here; use POST /super-admins
        password:
          type: string
          format: password
          minLength: 8
          description: >
            Set by the creating Org Admin (v1 has no email delivery). The user
            may change it later via POST /auth/change-password.
    BulkCreateParticipantsRequest:
      type: object
      required: [participants]
      properties:
        participants:
          type: array
          minItems: 1
          maxItems: 5000
          items:
            type: object
            required: [email, first_name, last_name]
            properties:
              email: { type: string, format: email }
              first_name: { type: string }
              last_name: { type: string }
    BulkCreateParticipantsResult:
      type: object
      properties:
        created_count: { type: integer }
        skipped_count: { type: integer }
        results:
          type: array
          items:
            type: object
            properties:
              email: { type: string, format: email }
              status: { type: string, enum: [CREATED, SKIPPED] }
              reason: { type: string, nullable: true, description: Present when SKIPPED (e.g. duplicate_email, invalid_email) }
              user_id: { type: string, format: uuid, nullable: true }
              one_time_password: { type: string, nullable: true, description: Returned only for CREATED rows; for out-of-band distribution }
    CreateSuperAdminRequest:
      type: object
      required: [email, first_name, last_name, password]
      properties:
        email: { type: string, format: email }
        first_name: { type: string }
        last_name: { type: string }
        password: { type: string, format: password, minLength: 8 }
...